+++ tests/all_tests.py
@@ -1,0 +1,9 @@
+
+
+from .test_email_code_verifier import Test_EmailCodeVerifier 
+from .test_email_template import TestEmailTemplate, TestFileEmailTemplate
+from .test_email_worker import Test_EmailWorker 
+from .test_gmail_sender import Test_send_Email_using_GmailSender
+from .test_preset_email_templates import *
+from .test_template_loading import  * 
+from .test_templatestyles_class import *