+++ RELEASE_NOTES.md
@@ -1,6 +1,18 @@
 # WHOLE MAIL RELEASE NOTES
 
-## VERSION v.0.1.7
+-------------------------------------------------------------------------------
+
+## VERSION 0.1.8
+
+- fixed the bug in sending gmail emails
+- Added *raise_exceptions* parameter to EmailSender
+
+    
+
+-------------------------------------------------------------------------------
+
+
+## VERSION 0.1.7
 
 - Temlates documentation and screenshots 
 - New templates - OfficeLetterEmailTemplate , InformativeMessageEmailTemplate , ResetPasswordLinkEmailTemplate